+++ tests/test_target_id_cpo.cpp
@@ -278,9 +278,9 @@ TEST_CASE("target_id(g,uv) - full graph traversal", "[target_id][cpo][integratio
         std::vector<std::pair<int, int>> all_edges;  // (source, target)
         
         for (auto u : vertices(graph)) {
-            int uid = vertex_id(graph, u);
+            int uid = static_cast<int>(vertex_id(graph, u));
             for (auto e : edges(graph, u)) {
-                int vid = target_id(graph, e);
+                int vid = static_cast<int>(target_id(graph, e));
                 all_edges.emplace_back(uid, vid);
             }
         }
